remove dead code & add analytics
