--- Stepic/ContentLanguagesView.xib
@@ -1,75 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="14113" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina4_7" orientation="portrait">
-        <adaptation id="fullscreen"/>
-    </device>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14088"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="ContentLanguagesView" customModule="Adaptive_1838" customModuleProvider="target">
-            <connections>
-                <outlet property="collectionView" destination="SaG-SC-tA2" id="Bls-Q0-spj"/>
-                <outlet property="descriptionLabel" destination="aps-oZ-dhi" id="BhI-Xz-WIG"/>
-                <outlet property="flowLayout" destination="1gR-IX-UOI" id="lOY-PR-lRh"/>
-                <outlet property="titleLabel" destination="2Df-Uu-cul" id="LSX-hR-yfT"/>
-            </connections>
-        </placeholder>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view contentMode="scaleToFill" id="iN0-l3-epB">
-            <rect key="frame" x="0.0" y="0.0" width="375" height="124"/>
-            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-            <subviews>
-                <collectionView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" dataMode="none" translatesAutoresizingMaskIntoConstraints="NO" id="SaG-SC-tA2">
-                    <rect key="frame" x="0.0" y="47" width="375" height="40"/>
-                    <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                    <constraints>
-                        <constraint firstAttribute="height" priority="999" constant="40" id="yLe-Fl-i6W"/>
-                    </constraints>
-                    <collectionViewFlowLayout key="collectionViewLayout" minimumLineSpacing="0.0" minimumInteritemSpacing="0.0" id="1gR-IX-UOI">
-                        <size key="itemSize" width="50" height="50"/>
-                        <size key="headerReferenceSize" width="0.0" height="0.0"/>
-                        <size key="footerReferenceSize" width="0.0" height="0.0"/>
-                        <inset key="sectionInset" minX="0.0" minY="0.0" maxX="0.0" maxY="0.0"/>
-                    </collectionViewFlowLayout>
-                </collectionView>
-                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Choose language" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="2Df-Uu-cul" customClass="StepikLabel" customModule="Adaptive_1838" customModuleProvider="target">
-                    <rect key="frame" x="0.0" y="0.0" width="375" height="27"/>
-                    <fontDescription key="fontDescription" type="system" pointSize="20"/>
-                    <nil key="textColor"/>
-                    <nil key="highlightedColor"/>
-                </label>
-                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="255" verticalCompressionResistancePriority="755" text="This can be also changed in preferences" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="aps-oZ-dhi" customClass="StepikLabel" customModule="Adaptive_1838" customModuleProvider="target">
-                    <rect key="frame" x="0.0" y="107" width="375" height="17"/>
-                    <fontDescription key="fontDescription" type="system" pointSize="14"/>
-                    <nil key="textColor"/>
-                    <nil key="highlightedColor"/>
-                    <userDefinedRuntimeAttributes>
-                        <userDefinedRuntimeAttribute type="boolean" keyPath="isGray" value="YES"/>
-                    </userDefinedRuntimeAttributes>
-                </label>
-            </subviews>
-            <color key="backgroundColor" red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-            <constraints>
-                <constraint firstItem="aps-oZ-dhi" firstAttribute="top" secondItem="SaG-SC-tA2" secondAttribute="bottom" constant="20" id="7BC-Em-et9"/>
-                <constraint firstItem="aps-oZ-dhi" firstAttribute="leading" secondItem="3BE-A2-TZ2" secondAttribute="leading" id="Onf-ou-T7Q"/>
-                <constraint firstItem="3BE-A2-TZ2" firstAttribute="trailing" secondItem="SaG-SC-tA2" secondAttribute="trailing" id="a5c-t6-Yun"/>
-                <constraint firstItem="3BE-A2-TZ2" firstAttribute="trailing" secondItem="2Df-Uu-cul" secondAttribute="trailing" id="aSZ-m0-n15"/>
-                <constraint firstItem="SaG-SC-tA2" firstAttribute="leading" secondItem="3BE-A2-TZ2" secondAttribute="leading" id="jhT-2q-AiQ"/>
-                <constraint firstItem="2Df-Uu-cul" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" id="nKQ-WU-M8t"/>
-                <constraint firstItem="2Df-Uu-cul" firstAttribute="leading" secondItem="3BE-A2-TZ2" secondAttribute="leading" id="qIg-tq-ivb"/>
-                <constraint firstItem="3BE-A2-TZ2" firstAttribute="trailing" secondItem="aps-oZ-dhi" secondAttribute="trailing" id="rTO-oJ-mjW"/>
-                <constraint firstItem="SaG-SC-tA2" firstAttribute="top" secondItem="2Df-Uu-cul" secondAttribute="bottom" constant="20" id="rgS-6k-Vfz"/>
-                <constraint firstItem="3BE-A2-TZ2" firstAttribute="bottom" secondItem="aps-oZ-dhi" secondAttribute="bottom" id="vNB-hI-n2f"/>
-            </constraints>
-            <nil key="simulatedTopBarMetrics"/>
-            <nil key="simulatedBottomBarMetrics"/>
-            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
-            <viewLayoutGuide key="safeArea" id="3BE-A2-TZ2"/>
-            <point key="canvasLocation" x="33.5" y="-181"/>
-        </view>
-    </objects>
-</document>
