--- Stepic/CourseTagsView.xib
@@ -1,62 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="14113" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina4_7" orientation="portrait">
-        <adaptation id="fullscreen"/>
-    </device>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14088"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="CourseTagsView" customModule="Adaptive_1838" customModuleProvider="target">
-            <connections>
-                <outlet property="collectionView" destination="b7J-Qq-Ns5" id="NPm-NY-dcg"/>
-                <outlet property="flowLayout" destination="0F3-tg-23g" id="UHv-FJ-gVE"/>
-                <outlet property="titleLabel" destination="VQK-z8-o1c" id="OmR-Cl-cQB"/>
-            </connections>
-        </placeholder>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view contentMode="scaleToFill" id="iN0-l3-epB">
-            <rect key="frame" x="0.0" y="0.0" width="375" height="81"/>
-            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-            <subviews>
-                <collectionView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" ambiguous="YES" showsHorizontalScrollIndicator="NO" showsVerticalScrollIndicator="NO" dataMode="none" translatesAutoresizingMaskIntoConstraints="NO" id="b7J-Qq-Ns5">
-                    <rect key="frame" x="0.0" y="44" width="375" height="37"/>
-                    <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                    <constraints>
-                        <constraint firstAttribute="height" priority="750" constant="40" id="DHw-Bg-PGm"/>
-                    </constraints>
-                    <collectionViewFlowLayout key="collectionViewLayout" scrollDirection="horizontal" minimumLineSpacing="10" minimumInteritemSpacing="10" id="0F3-tg-23g">
-                        <size key="itemSize" width="50" height="50"/>
-                        <size key="headerReferenceSize" width="0.0" height="0.0"/>
-                        <size key="footerReferenceSize" width="0.0" height="0.0"/>
-                        <inset key="sectionInset" minX="16" minY="0.0" maxX="16" maxY="0.0"/>
-                    </collectionViewFlowLayout>
-                </collectionView>
-                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" ambiguous="YES" text="Trending topics" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="VQK-z8-o1c" customClass="StepikLabel" customModule="Adaptive_1838" customModuleProvider="target">
-                    <rect key="frame" x="16" y="0.0" width="333" height="24"/>
-                    <fontDescription key="fontDescription" type="system" pointSize="20"/>
-                    <nil key="textColor"/>
-                    <nil key="highlightedColor"/>
-                </label>
-            </subviews>
-            <color key="backgroundColor" red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-            <constraints>
-                <constraint firstItem="VUZ-6F-fqi" firstAttribute="trailing" secondItem="b7J-Qq-Ns5" secondAttribute="trailing" id="5n6-FI-uhT"/>
-                <constraint firstItem="VQK-z8-o1c" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" id="Jmk-hz-fQH"/>
-                <constraint firstItem="b7J-Qq-Ns5" firstAttribute="top" secondItem="VQK-z8-o1c" secondAttribute="bottom" constant="20" id="b8C-wQ-Cc0"/>
-                <constraint firstItem="b7J-Qq-Ns5" firstAttribute="leading" secondItem="VUZ-6F-fqi" secondAttribute="leading" id="dtW-ez-PS5"/>
-                <constraint firstItem="VUZ-6F-fqi" firstAttribute="trailing" secondItem="VQK-z8-o1c" secondAttribute="trailing" constant="26" id="e2P-Eb-foJ"/>
-                <constraint firstAttribute="bottom" secondItem="b7J-Qq-Ns5" secondAttribute="bottom" id="eFO-dN-WHr"/>
-                <constraint firstItem="VQK-z8-o1c" firstAttribute="leading" secondItem="VUZ-6F-fqi" secondAttribute="leading" constant="16" id="lhu-pn-l6S"/>
-            </constraints>
-            <nil key="simulatedTopBarMetrics"/>
-            <nil key="simulatedBottomBarMetrics"/>
-            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
-            <viewLayoutGuide key="safeArea" id="VUZ-6F-fqi"/>
-            <point key="canvasLocation" x="33.5" y="-77.5"/>
-        </view>
-    </objects>
-</document>
