--- Stepic/CourseWidgetCollectionViewCell.xib
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="14113" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <dependencies>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14088"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <collectionViewCell opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" id="gTV-IL-0wX" customClass="CourseWidgetCollectionViewCell" customModule="Adaptive_1838" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="202" height="76"/>
-            <autoresizingMask key="autoresizingMask"/>
-            <view key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center">
-                <rect key="frame" x="0.0" y="0.0" width="202" height="76"/>
-                <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-                <subviews>
-                    <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="UHe-Fq-AdG" customClass="OldCourseWidgetView" customModule="Adaptive_1838" customModuleProvider="target">
-                        <rect key="frame" x="0.0" y="0.0" width="202" height="76"/>
-                        <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                        <viewLayoutGuide key="safeArea" id="LJZ-Gh-YCE"/>
-                    </view>
-                </subviews>
-            </view>
-            <constraints>
-                <constraint firstAttribute="trailing" secondItem="UHe-Fq-AdG" secondAttribute="trailing" priority="999" id="B1O-n6-lo0"/>
-                <constraint firstAttribute="bottom" secondItem="UHe-Fq-AdG" secondAttribute="bottom" priority="999" id="CaO-Y4-EA8"/>
-                <constraint firstItem="UHe-Fq-AdG" firstAttribute="leading" secondItem="gTV-IL-0wX" secondAttribute="leading" id="Vyf-6H-hvP"/>
-                <constraint firstItem="UHe-Fq-AdG" firstAttribute="top" secondItem="gTV-IL-0wX" secondAttribute="top" id="fy9-bk-OHA"/>
-            </constraints>
-            <viewLayoutGuide key="safeArea" id="4Hc-QT-ab0"/>
-            <size key="customSize" width="202" height="76"/>
-            <connections>
-                <outlet property="widgetView" destination="UHe-Fq-AdG" id="cLM-aS-hoj"/>
-            </connections>
-            <point key="canvasLocation" x="110" y="67"/>
-        </collectionViewCell>
-    </objects>
-</document>
