--- Stepic/UserActivityHomeView.xib
@@ -1,68 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="14109" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina4_7" orientation="portrait">
-        <adaptation id="fullscreen"/>
-    </device>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14088"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="UserActivityHomeView" customModule="Adaptive_1838" customModuleProvider="target">
-            <connections>
-                <outlet property="streakCountLabel" destination="wgg-D3-kRH" id="z62-nE-j6W"/>
-                <outlet property="streakTextLabel" destination="ROf-Sa-jjo" id="mHh-oR-Sew"/>
-            </connections>
-        </placeholder>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view contentMode="scaleToFill" insetsLayoutMarginsFromSafeArea="NO" id="iN0-l3-epB">
-            <rect key="frame" x="0.0" y="0.0" width="316" height="60"/>
-            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-            <subviews>
-                <imageView userInteractionEnabled="NO" contentMode="scaleToFill" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="lightning_dark" translatesAutoresizingMaskIntoConstraints="NO" id="MmC-2s-9gC">
-                    <rect key="frame" x="4" y="22.5" width="15" height="15"/>
-                    <constraints>
-                        <constraint firstAttribute="width" constant="15" id="CNB-kj-D9U"/>
-                        <constraint firstAttribute="height" constant="15" id="lbl-eh-yIC"/>
-                    </constraints>
-                </imageView>
-                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" horizontalCompressionResistancePriority="751" text="10" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="wgg-D3-kRH" customClass="StepikLabel" customModule="Adaptive_1838" customModuleProvider="target">
-                    <rect key="frame" x="20" y="20" width="18" height="20"/>
-                    <fontDescription key="fontDescription" type="system" weight="medium" pointSize="16"/>
-                    <nil key="textColor"/>
-                    <nil key="highlightedColor"/>
-                </label>
-                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="249" verticalHuggingPriority="240" verticalCompressionResistancePriority="760" text="Your current streak is N days! Great" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="ROf-Sa-jjo" customClass="StepikLabel" customModule="Adaptive_1838" customModuleProvider="target">
-                    <rect key="frame" x="58" y="8" width="250" height="44"/>
-                    <constraints>
-                        <constraint firstAttribute="height" relation="greaterThanOrEqual" constant="16" id="Rhk-H9-DFQ"/>
-                    </constraints>
-                    <fontDescription key="fontDescription" type="system" weight="light" pointSize="16"/>
-                    <nil key="textColor"/>
-                    <nil key="highlightedColor"/>
-                </label>
-            </subviews>
-            <color key="backgroundColor" red="0.27058823529411763" green="0.69019607843137254" blue="1" alpha="0.080000000000000002" colorSpace="custom" customColorSpace="sRGB"/>
-            <constraints>
-                <constraint firstItem="MmC-2s-9gC" firstAttribute="leading" secondItem="iN0-l3-epB" secondAttribute="leading" constant="4" id="DfP-1V-Rwz"/>
-                <constraint firstItem="MmC-2s-9gC" firstAttribute="centerY" secondItem="iN0-l3-epB" secondAttribute="centerY" id="G0Z-19-TGp"/>
-                <constraint firstItem="wgg-D3-kRH" firstAttribute="centerY" secondItem="iN0-l3-epB" secondAttribute="centerY" id="Jk1-zK-efe"/>
-                <constraint firstItem="wgg-D3-kRH" firstAttribute="leading" secondItem="MmC-2s-9gC" secondAttribute="trailing" constant="1" id="MJE-gf-P71"/>
-                <constraint firstItem="ROf-Sa-jjo" firstAttribute="leading" secondItem="wgg-D3-kRH" secondAttribute="trailing" constant="20" id="QK8-Kq-lFt"/>
-                <constraint firstAttribute="bottom" secondItem="ROf-Sa-jjo" secondAttribute="bottom" constant="8" id="X7e-1t-BdD"/>
-                <constraint firstAttribute="trailing" secondItem="ROf-Sa-jjo" secondAttribute="trailing" constant="8" id="Zw1-Go-dxW"/>
-                <constraint firstItem="ROf-Sa-jjo" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" constant="8" id="cS3-iZ-sa8"/>
-            </constraints>
-            <nil key="simulatedTopBarMetrics"/>
-            <nil key="simulatedBottomBarMetrics"/>
-            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
-            <viewLayoutGuide key="safeArea" id="zUw-Sz-eeh"/>
-            <point key="canvasLocation" x="-5" y="145"/>
-        </view>
-    </objects>
-    <resources>
-        <image name="lightning_dark" width="20" height="20"/>
-    </resources>
-</document>
